WIP pass single color to span color callback
We don't have visible merged spans anymore. The only reason to pass a
slice of spans is because of our metadata abstraction. However, calling
Slice allocates (putting structs in interface values), and none of our
color functions care about metadata.

In the future we'll need to figure out a better way to have per-span metadata.